requirements.txt
# SMAIT HRI v3.0 Dependencies
# =====================================
# Install: pip install -r requirements.txt

# Core
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
websockets>=12.0

# Audio
sounddevice>=0.4.6

# VAD (Silero VAD via torch.hub)
torch>=2.0.0
torchaudio>=2.0.0

# Speech Separation (Dolphin AV-TSE)
# Install from: pip install git+https://github.com/JusperLee/Dolphin.git
einops

# ASR (NVIDIA Parakeet TDT 0.6B v2)
nemo_toolkit[asr]>=2.0.0

# TTS (Kokoro-82M)
# Install from: pip install kokoro  (or from HuggingFace: hexgrad/Kokoro-82M)

# Face tracking + Gaze
mediapipe>=0.10.0
# L2CS-Net: pip install git+https://github.com/Ahmednull/L2CS-Net.git

# EOU detection (LiveKit turn detector)
# Install from: pip install git+https://github.com/livekit/turn-detector.git
transformers>=4.35.0

# LLM (Ollama local + OpenAI API fallback)
openai>=1.0.0
requests>=2.31.0

# Async performance
uvloop>=0.19.0; sys_platform != 'win32'

# System monitoring
psutil
